Refactor utils into common module
