+++ mlflow/sagemaker/__init__.py
@@ -12,18 +12,18 @@ import time
 
 import mlflow
 import mlflow.version
-from mlflow import pyfunc, mleap
 from mlflow.exceptions import MlflowException
 from mlflow.models import Model
-from mlflow.protos.databricks_pb2 import RESOURCE_DOES_NOT_EXIST, INVALID_PARAMETER_VALUE
+from mlflow.models.docker_utils import DEFAULT_IMAGE_NAME, _get_preferred_deployment_flavor,\
+    _validate_deployment_flavor
+from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
 from mlflow.tracking.artifact_utils import _download_artifact_from_uri
 from mlflow.utils import get_unique_resource_id
-from mlflow.utils.file_utils import TempDir, _copy_project
+from mlflow.utils.file_utils import TempDir
 from mlflow.utils.logging_utils import eprint
 from mlflow.sagemaker.container import SUPPORTED_FLAVORS as SUPPORTED_DEPLOYMENT_FLAVORS
-from mlflow.sagemaker.container import DEPLOYMENT_CONFIG_KEY_FLAVOR_NAME, DISABLE_ENV_CREATION
+from mlflow.sagemaker.container import DEPLOYMENT_CONFIG_KEY_FLAVOR_NAME
 
-DEFAULT_IMAGE_NAME = "mlflow-pyfunc"
 
 DEPLOYMENT_MODE_ADD = "add"
 DEPLOYMENT_MODE_REPLACE = "replace"
@@ -44,142 +44,7 @@ DEFAULT_BUCKET_NAME_PREFIX = "mlflow-sagemaker"
 DEFAULT_SAGEMAKER_INSTANCE_TYPE = "ml.m4.xlarge"
 DEFAULT_SAGEMAKER_INSTANCE_COUNT = 1
 
-_DOCKERFILE_TEMPLATE = """
-# Build an image that can serve pyfunc model in SageMaker
-FROM ubuntu:16.04
-
-RUN apt-get -y update && apt-get install -y --no-install-recommends \
-         wget \
-         curl \
-         nginx \
-         ca-certificates \
-         bzip2 \
-         build-essential \
-         cmake \
-         openjdk-8-jdk \
-         git-core \
-         maven \
-    && rm -rf /var/lib/apt/lists/*
-
-# Download and setup miniconda
-RUN curl https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh >> miniconda.sh
-RUN bash ./miniconda.sh -b -p /miniconda; rm ./miniconda.sh;
-ENV PATH="/miniconda/bin:$PATH"
-ENV JAVA_HOME=/usr/lib/jvm/java-8-openjdk-amd64
-
-RUN conda install gunicorn;\
-    conda install gevent;\
-
-{install_mlflow}
-{setup_model}
-
-# Set up the program in the image
-WORKDIR /opt/mlflow
-
-# start mlflow scoring
-ENTRYPOINT ["python", "-c", "import sys; from mlflow.sagemaker import container as C; \
-C._init(sys.argv[1])"]
-"""
-
 _logger = logging.getLogger(__name__)
-
-
-def _docker_ignore(mlflow_root):
-    docker_ignore = os.path.join(mlflow_root, '.dockerignore')
-
-    def strip_slash(x):
-        if x.startswith("/"):
-            x = x[1:]
-        if x.endswith('/'):
-            x = x[:-1]
-        return x
-
-    if os.path.exists(docker_ignore):
-        with open(docker_ignore, "r") as f:
-            patterns = [x.strip() for x in f.readlines()]
-            patterns = [strip_slash(x)
-                        for x in patterns if not x.startswith("#")]
-
-    def ignore(_, names):
-        import fnmatch
-        res = set()
-        for p in patterns:
-            res.update(set(fnmatch.filter(names, p)))
-        return list(res)
-
-    return ignore
-
-
-def build_image(name=DEFAULT_IMAGE_NAME, mlflow_home=None, model_uri=None, flavor=None):
-    """
-    Build an MLflow Docker image.
-    The image is built locally and it requires Docker to run.
-
-    :param name: Docker image name.
-    :param mlflow_home: (Optional) Path to a local copy of the MLflow GitHub repository.
-                        If specified, the image will install MLflow from this directory.
-                        If None, it will install MLflow from pip.
-    """
-    with TempDir() as tmp:
-        cwd = tmp.path()
-        if mlflow_home:
-            mlflow_dir = _copy_project(
-                src_path=mlflow_home, dst_path=cwd)
-            install_mlflow = (
-                "COPY {mlflow_dir} /opt/mlflow\n"
-                "RUN pip install /opt/mlflow\n"
-            ).format(mlflow_dir=mlflow_dir)
-        else:
-            install_mlflow = (
-                "RUN pip install mlflow=={version}\n"
-                "RUN mvn --batch-mode dependency:copy"
-                " -Dartifact=org.mlflow:mlflow-scoring:{version}:pom"
-                " -DoutputDirectory=/opt/java\n"
-                "RUN mvn --batch-mode dependency:copy"
-                " -Dartifact=org.mlflow:mlflow-scoring:{version}:jar"
-                " -DoutputDirectory=/opt/java/jars\n"
-                "RUN cd /opt/java && mv mlflow-scoring-{version}.pom pom.xml &&"
-                " mvn --batch-mode dependency:copy-dependencies -DoutputDirectory=/opt/java/jars\n"
-                "RUN rm /opt/java/pom.xml"
-            ).format(version=mlflow.version.VERSION)
-        if model_uri:
-            model_cwd = tmp.path("model_dir")
-            os.mkdir(model_cwd)
-            model_path = _download_artifact_from_uri(model_uri, output_path=model_cwd)
-            model_config_path = os.path.join(model_path, "MLmodel")
-            model_config = Model.load(model_config_path)
-
-            if flavor is None:
-                flavor = _get_preferred_deployment_flavor(model_config)
-            else:
-                _validate_deployment_flavor(model_config, flavor)
-            print("Using the {selected_flavor} flavor for local "
-                  "serving!".format(selected_flavor=flavor))
-            copy_model_into_container = ("""
-            ENV {disable_env}="true"
-            COPY {model_dir} /opt/ml/model
-            RUN python -c \
-            'from mlflow.sagemaker.container import _install_base_deps;\
-            _install_base_deps("/opt/ml/model")'
-            """.format(disable_env=DISABLE_ENV_CREATION,
-                       model_dir=os.path.join("model_dir", os.path.basename(model_path)))
-            )
-        else:
-            copy_model_into_container = ""
-
-        with open(os.path.join(cwd, "Dockerfile"), "w") as f:
-            f.write(_DOCKERFILE_TEMPLATE.format(
-                install_mlflow=install_mlflow, setup_model=copy_model_into_container))
-        _logger.info("building docker image")
-        os.system('find {cwd}/'.format(cwd=cwd))
-        proc = Popen(["docker", "build", "-t", name, "-f", "Dockerfile", "."],
-                     cwd=cwd,
-                     stdout=PIPE,
-                     stderr=STDOUT,
-                     universal_newlines=True)
-        for x in iter(proc.stdout.readline, ""):
-            eprint(x, end='')
-
 
 _full_template = "{account}.dkr.ecr.{region}.amazonaws.com/{image}:{version}"
 
@@ -419,56 +284,6 @@ def deploy(app_name, model_uri, execution_role_arn=None, bucket=None,
                 " \"{error_message}\"".format(error_message=operation_status.message))
         if not archive:
             deployment_operation.clean_up()
-
-
-def _get_preferred_deployment_flavor(model_config):
-    """
-    Obtains the flavor that MLflow would prefer to use when deploying the model.
-    If the model does not contain any supported flavors for deployment, an exception
-    will be thrown.
-
-    :param model_config: An MLflow model object
-    :return: The name of the preferred deployment flavor for the specified model
-    """
-    if mleap.FLAVOR_NAME in model_config.flavors:
-        return mleap.FLAVOR_NAME
-    elif pyfunc.FLAVOR_NAME in model_config.flavors:
-        return pyfunc.FLAVOR_NAME
-    else:
-        raise MlflowException(
-                message=(
-                    "The specified model does not contain any of the supported flavors for"
-                    " deployment. The model contains the following flavors: {model_flavors}."
-                    " Supported flavors: {supported_flavors}".format(
-                        model_flavors=model_config.flavors.keys(),
-                        supported_flavors=SUPPORTED_DEPLOYMENT_FLAVORS)),
-                error_code=RESOURCE_DOES_NOT_EXIST)
-
-
-def _validate_deployment_flavor(model_config, flavor):
-    """
-    Checks that the specified flavor is a supported deployment flavor
-    and is contained in the specified model. If one of these conditions
-    is not met, an exception is thrown.
-
-    :param model_config: An MLflow Model object
-    :param flavor: The deployment flavor to validate
-    """
-    if flavor not in SUPPORTED_DEPLOYMENT_FLAVORS:
-        raise MlflowException(
-                message=(
-                    "The specified flavor: `{flavor_name}` is not supported for deployment."
-                    " Please use one of the supported flavors: {supported_flavor_names}".format(
-                        flavor_name=flavor,
-                        supported_flavor_names=SUPPORTED_DEPLOYMENT_FLAVORS)),
-                error_code=INVALID_PARAMETER_VALUE)
-    elif flavor not in model_config.flavors:
-        raise MlflowException(
-                message=("The specified model does not contain the specified deployment flavor:"
-                         " `{flavor_name}`. Please use one of the following deployment flavors"
-                         " that the model contains: {model_flavors}".format(
-                             flavor_name=flavor, model_flavors=model_config.flavors.keys())),
-                error_code=RESOURCE_DOES_NOT_EXIST)
 
 
 def delete(app_name, region_name="us-west-2", archive=False, synchronous=True, timeout_seconds=300):
